+++ test/ccsave.test.ts
@@ -2419,7 +2419,7 @@ test.describe('CCSave.toNativeSave edge cases:', () => {
 
     test('Empty pantheon slots parse from -1 and back', async ({ browser }) => {
         let page = await openCookieClickerPage(browser);
-        await page.evaluate( () => CConnoisseur.gainLumps(1) ); // Unlock lumps
+        await page.evaluate( () => Game.lumps = Game.lumpsTotal = 1 ); // Unlock lumps
         await page.evaluate( () => Game.Objects['Temple'].getFree(1) ); // Get a temple
         await page.evaluate( () => Game.Objects['Temple'].levelUp() ); // Unlock Pantheon
         // We must wait for the minigame to load before Game.WriteSave.